requirements.txt
apispec==5.2.2
attrs==22.1.0
click==8.1.3
colorama==0.4.5
flasgger==0.9.5
Flask==2.1.3
flask-marshmallow==0.14.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.8.0
MarkupSafe==2.1.1
marshmallow==3.17.0
mistune==2.0.4
numpy==1.23.3
packaging==21.3
pandas==1.4.4
pyparsing==3.0.9
pyrsistent==0.18.1
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.2.1
PyYAML==6.0
scipy==1.9.1
six==1.16.0
waitress==2.1.2
Werkzeug==2.2.1
